+++ src/pages/admin/comments.tsx
@@ -1,24 +1,21 @@
 import React from 'react';
 import { GetServerSideProps } from 'next';
-import { useMutation, useQueryClient } from 'react-query';
 
 import { useCommentListQuery } from '../../api/hook/admin';
-import { deleteComment } from '../../api/fetcher/admin';
 import { useIsValidAdmin } from '../../hooks/useIsValidAdmin';
 
 import HomeLayout from '../../components/common/layout/layout';
-import { Container, Title, Section, Tr, Td, PageContent } from '../../styles/admin/user';
 import Button from '../../components/common/button';
 import ArrowLeftIcon from '../../components/common/@Icons/arrowLeftIcon';
 import ArrowRightIcon from '../../components/common/@Icons/arrowRightIcon';
-import GarbageIcon from '../../components/common/@Icons/garbageIcon';
+import { Container, Title, Section, PageContent } from '../../styles/admin/user';
+import DataListTable from '../../components/admin/DataListTable';
 
 const Comments = () => {
   const [page, setPage] = React.useState<number>(1);
   const [totalPage, setTotalPage] = React.useState<number>(0);
   const isAdmin = useIsValidAdmin();
 
-  const queryClient = useQueryClient();
   const { data: commentList } = useCommentListQuery({
     per_page: 10,
     page,
@@ -30,36 +27,11 @@ const Comments = () => {
     },
   });
 
-  const { mutate: deleteCommentMutation } = useMutation(deleteComment, {
-    onSuccess: () => {
-      queryClient.invalidateQueries(['COMMENTSLIST', page]);
-    },
-  });
-
   return (
     <Container>
       <Title>어드민페이지_댓글관리</Title>
       <Section>
-        <table>
-          <thead>
-            <Tr header>
-              <th>닉네임</th>
-              <th>내용</th>
-              <th>댓글삭제</th>
-            </Tr>
-          </thead>
-          <tbody>
-            {commentList?.rows?.map(data => (
-              <tr key={data.id}>
-                <Td>{data?.User?.nickname}</Td>
-                <Td>{data.content}</Td>
-                <Td>
-                  <GarbageIcon onClick={() => deleteCommentMutation(data.id)} />
-                </Td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
+        <DataListTable commentList={commentList} th_1="닉네임" th_2="내용" th_3="댓글삭제" page={page} />
       </Section>
 
       <PageContent>
