+++ src/pages/admin/user.tsx
@@ -1,26 +1,22 @@
 import React from 'react';
 import { GetServerSideProps } from 'next';
-import { useMutation, useQueryClient } from 'react-query';
 
 import { useUserListQuery } from '../../api/hook/admin';
-import { dateFormat } from '../../utils/dateFormat';
-import { deleteUser } from '../../api/fetcher/admin';
 import { useIsValidAdmin } from '../../hooks/useIsValidAdmin';
 
+import HomeLayout from '../../components/common/layout/layout';
 import Button from '../../components/common/button';
 import ArrowLeftIcon from '../../components/common/@Icons/arrowLeftIcon';
 import ArrowRightIcon from '../../components/common/@Icons/arrowRightIcon';
-import HomeLayout from '../../components/common/layout/layout';
-import { Container, Title, Section, Tr, PageContent, Td } from '../../styles/admin/user';
-import GarbageIcon from '../../components/common/@Icons/garbageIcon';
+import { Container, Title, Section, PageContent } from '../../styles/admin/user';
+import DataListTable from '../../components/admin/DataListTable';
 
 const Users = () => {
   const [page, setPage] = React.useState<number>(1);
   const [totalPage, setTotalPage] = React.useState<number>(0);
   const isAdmin = useIsValidAdmin();
 
-  const queryClient = useQueryClient();
-  const { data } = useUserListQuery({
+  const { data: userList } = useUserListQuery({
     per_page: 10,
     page,
     options: {
@@ -31,40 +27,11 @@ const Users = () => {
     },
   });
 
-  const { mutate: deleteUserMutation } = useMutation(deleteUser, {
-    onSuccess: () => {
-      queryClient.invalidateQueries(['USERLIST', page]);
-    },
-  });
-
-  const handleUserWithdraw = (id: number) => {
-    deleteUserMutation(id);
-  };
-
   return (
     <Container>
       <Title>어드민페이지_유저관리</Title>
       <Section>
-        <table>
-          <thead>
-            <Tr header>
-              <th>닉네임</th>
-              <th>가입일자</th>
-              <th>유저탈퇴</th>
-            </Tr>
-          </thead>
-          <tbody>
-            {data?.rows?.map(data => (
-              <tr key={data.id}>
-                <Td>{data.nickname}</Td>
-                <Td>{dateFormat(data.createdAt)}</Td>
-                <Td>
-                  <GarbageIcon onClick={() => handleUserWithdraw(data.id)} />
-                </Td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
+        <DataListTable userList={userList} th_1="닉네임" th_2="가입일자" th_3="유저탈퇴" page={page} />
       </Section>
       <PageContent>
         <Button onClick={() => setPage(old => Math.max(old - 1, 1))}>
